Restart sync if delta link fails with 401
Restart the sync from the initial link when a 401 is received. This
is required when an old delta link is used, even though the
authorization token is valid.

Remove the reset of the error count during handling of a reset. This is
mostly to prevent a real 401 (from an invalid authorization token) from
looping indefinitely, but also removes an infinite loop if the API
decides to return 410 repeatedly.

Fixes #1